Feat: Implement new cli flag --update-test-actions
Modified default behaviour; action calls containing "test" in the comment string on
the same line will no longer be modified by default. This is a critical piece of
functionality where repositories have action calls that are on branches where
testing is being performed prior to a new tag/release of the action.

Signed-off-by: Matthew Watkins <mwatkins@linuxfoundation.org>